pyproject.toml
[project]
name = "dailyassistant"
version = "0.1.0"
description = "DailyAssistant automation toolkit (Google Docs/Slides/Drive, email workflows, OpenRouter LLM helpers, PDF utilities)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [ { name = "DailyAssistant" } ]
dependencies = [
	"pytest>=8.2.0",
	"faiss-cpu>=1.7.4",
	"python-frontmatter>=1.0.0",
	"numpy>=1.26.0",
	"streamlit>=1.33.0",
	"pytest-cov>=5.0.0",
	"ruff>=0.5.0",
]

[project.optional-dependencies]
google = ["google-api-python-client", "google-auth", "google-auth-oauthlib"]
email = ["beautifulsoup4"]
pdf = ["markdown", "pdfminer.six"]
openrouter = []
all = ["google-api-python-client", "google-auth", "google-auth-oauthlib", "beautifulsoup4", "markdown", "pdfminer.six"]

[project.scripts]
da = "dailyassistant.cli:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tools", "dailyassistant"]

[tool.setuptools.package-dir]
tools = "tools"
dailyassistant = "dailyassistant"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["dailyassistant", "tools"]

[tool.coverage.run]
branch = true
source = ["projects/ZoteroPDF/scripts", "projects/ZoteroPDF/openRouterAI"]
omit = ["*/__init__.py", "*/tests/*"]

[tool.coverage.report]
skip_empty = true
show_missing = true
